# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Mosaic API
  version: 1.0.0
  description: |
    Zero-knowledge encrypted photo gallery API.
    
    The server never sees plaintext photos or metadata—all encryption/decryption 
    happens client-side. The server stores encrypted blobs and facilitates 
    synchronization between clients.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: API base path (behind reverse proxy)
security:
  - RemoteUser: []
...
paths:
...
  /albums/{albumId}/epoch-keys/{keyId}:
    get:
      operationId: getEpochKey
      summary: Get one epoch key bundle
      description: Returns a single encrypted epoch key bundle for its recipient.
      tags: [Epoch Keys]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - name: keyId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Epoch key record
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EpochKeyRecord' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

  # ===========================================================================
  # Members
  # ===========================================================================
  /albums/{albumId}/members:
    get:
      operationId: listAlbumMembers
      summary: List album members
      tags: [Members]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - $ref: '#/components/parameters/PageSkip'
        - $ref: '#/components/parameters/PageTake'
      responses:
        '200':
          description: List of members
          headers:
            X-Pagination-Skip:
              $ref: '#/components/headers/PaginationSkip'
            X-Pagination-Take:
              $ref: '#/components/headers/PaginationTake'
            X-Pagination-Total-Count:
              $ref: '#/components/headers/PaginationTotalCount'
            X-Pagination-Has-More:
              $ref: '#/components/headers/PaginationHasMore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedAlbumMembers'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: inviteToAlbum
      summary: Invite a user to the album
      description: |
        Creates membership and distributes epoch keys to the new member.
        The epoch keys array must contain sealed bundles for all current epochs.
      tags: [Members]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InviteRequest'
      responses:
        '201':
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AlbumMember'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /albums/{albumId}/members/{userId}:
    delete:
      operationId: removeAlbumMember
      summary: Remove a member from the album
      description: |
        Revokes membership. A new epoch should be created after removal
        to ensure post-compromise security — removed members cannot decrypt content
        uploaded under epochs they were not a part of.
      tags: [Members]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: Member removed
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ===========================================================================
  # Epoch Keys
  # ===========================================================================
  /albums/{albumId}/epoch-keys:
    get:
      operationId: getEpochKeys
      summary: Get epoch keys for current user
      description: Returns all epoch key bundles encrypted for the current user.
      tags: [Epoch Keys]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: List of epoch key records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EpochKeyRecord'
        '403':
          $ref: '#/components/responses/Forbidden'

    post:
      operationId: createEpochKey
      summary: Create a new epoch key record
      description: |
        Used when rotating keys or sharing with new members.
        The encrypted bundle is sealed to the recipient.
      tags: [Epoch Keys]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEpochKeyRequest'
      responses:
        '201':
          description: Epoch key created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EpochKeyRecord'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
  /albums/{albumId}/epochs/{epochId}/rotate:
    post:
      operationId: rotateEpoch
      summary: Create a new epoch (key rotation)
      description: |
        Creates a new epoch with fresh keys. Used after member removal.
        Must include epoch keys for all current members.
      tags: [Epoch Keys]
      parameters:
        - $ref: '#/components/parameters/AlbumId'
        - name: epochId
          in: path
          required: true
          description: New epoch ID (must be greater than current)
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RotateEpochRequest'
      responses:
        '201':
          description: Epoch created
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'

  # ===========================================================================
  # Manifests
  # ===========================================================================
...
components:
...
  securitySchemes:
    RemoteUser:
      type: apiKey
      in: header
      name: Remote-User
      description: |
        User identity from trusted reverse proxy.
        Only accepted from configured trusted IP ranges.
    cookieAuth:
      type: apiKey
      in: cookie
      name: mosaic_session
      description: |
        LocalAuth session cookie. The backend sets this cookie as HttpOnly with Path=/api,
        SameSite=Lax for development/testing HTTP, and Secure + SameSite=Strict in production.

  # ===========================================================================
  # Parameters
  # ===========================================================================
  parameters:
    AlbumId:
      name: albumId
      in: path
      required: true
      description: Album UUID
      schema:
        type: string
        format: uuid
    UserId:
      name: userId
      in: path
      required: true
      description: User UUID
      schema:
        type: string
        format: uuid
    PageSkip:
      name: skip
      in: query
      required: false
      description: Number of rows to skip before returning a page.
      schema:
        type: integer
        minimum: 0
        default: 0
    PageTake:
      name: take
      in: query
      required: false
      description: Page size. Values are clamped to the backend maximum of 100.
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
...
  headers:
    PaginationSkip:
      description: Normalized skip value applied by the backend.
      schema:
        type: integer
    PaginationTake:
      description: Normalized take value applied by the backend.
      schema:
        type: integer
    PaginationTotalCount:
      description: Total rows available for the query before paging.
      schema:
        type: integer
    PaginationHasMore:
      description: Whether another page exists after this response.
      schema:
        type: boolean

  # ===========================================================================
  # Responses
  # ===========================================================================
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    Forbidden:
      description: Insufficient permissions
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807 problem details error response.
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
          description: Request correlation identifier when available.
      additionalProperties: true
    PagedResult:
      type: object
      required: [items, nextSkip]
      properties:
        items:
          type: array
          items: {}
        nextSkip:
          type: integer
          nullable: true
          description: Skip value for the next page, or null when no further page is available.
...
    PagedAlbumMembers:
      allOf:
        - $ref: '#/components/schemas/PagedResult'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/AlbumMember'
...
    UserPublic:
      type: object
      required:
        - id
        - identityPubkey
      properties:
        id:
          type: string
          format: uuid
        identityPubkey:
          type: string
          description: Base64-encoded Ed25519 public key
...
    AlbumMember:
      type: object
      required:
        - userId
        - role
        - joinedAt
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, editor, viewer]
        invitedBy:
          type: string
          format: uuid
        joinedAt:
          type: string
          format: date-time
        user:
          $ref: '#/components/schemas/UserPublic'
          description: Included user info for display
    InviteRequest:
      type: object
      required:
        - recipientId
        - role
        - epochKeys
      properties:
        recipientId:
          type: string
          format: uuid
          description: User to invite
        role:
          type: string
          enum: [editor, viewer]
        epochKeys:
          type: array
          items:
            $ref: '#/components/schemas/CreateEpochKeyRequest'
          description: Sealed epoch key bundles for all epochs

    # -------------------------------------------------------------------------
    # Epoch Keys
    # -------------------------------------------------------------------------
    EpochKeyRecord:
      type: object
      required:
        - id
        - albumId
        - epochId
        - encryptedKeyBundle
        - ownerSignature
        - sharerPubkey
        - signPubkey
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        albumId:
          type: string
          format: uuid
        epochId:
          type: integer
        encryptedKeyBundle:
          type: string
          format: byte
          description: Base64-encoded sealed box ciphertext
        ownerSignature:
          type: string
          format: byte
          description: Base64-encoded Ed25519 signature
        sharerPubkey:
          type: string
          format: byte
          description: Base64-encoded Ed25519 public key of sharer
        signPubkey:
          type: string
          format: byte
          description: Base64-encoded epoch signing public key
        createdAt:
          type: string
          format: date-time
    CreateEpochKeyRequest:
      type: object
      required:
        - recipientId
        - epochId
        - encryptedKeyBundle
        - ownerSignature
        - sharerPubkey
        - signPubkey
      properties:
        recipientId:
          type: string
          format: uuid
        epochId:
          type: integer
        encryptedKeyBundle:
          type: string
          format: byte
          description: Base64-encoded sealed box ciphertext
        ownerSignature:
          type: string
          format: byte
          description: Base64-encoded Ed25519 signature
        sharerPubkey:
          type: string
          format: byte
          description: Base64-encoded Ed25519 public key of sharer
        signPubkey:
          type: string
          format: byte
          description: Base64-encoded epoch signing public key
    RotateEpochRequest:
      type: object
      required:
        - epochKeys
      properties:
        epochKeys:
          type: array
          items:
            $ref: '#/components/schemas/CreateEpochKeyRequest'
          description: Epoch key bundles for all current members

    # -------------------------------------------------------------------------
    # Sync
    # -------------------------------------------------------------------------
...